requirements.txt
dash
dash-bootstrap-components
psycopg2-binary
python-dotenv
gunicorn
cryptography
Flask
dash_auth
